FESI1-40 [Feat]: 드롭다운 외부 클릭 시 자동 닫힘 & 컴포넌트 정리
