+++ src/components/@shared/dropdown/Dropdown.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useState } from 'react';
+import { useState, useEffect, useRef } from 'react';
 import { clsx } from 'clsx';
 import Image from 'next/image';
 
@@ -13,59 +13,111 @@ interface DropdownProps {
   // 드롭다운에 표시할 옵션들의 배열
   options: DropdownOption[];
   // 드롭다운 초기 값
-  defaultValue?: DropdownOption;
+  defaultLabel: string;
   // 옵션 선택 시 호출되는 콜백 함수
   onChange?: (option: DropdownOption) => void;
   // 추가 스타일링 클래스
   className?: string;
+  icon?: string;
+  selectedValue?: string;
 }
 
 /**
  * 지역 선택을 위한 드롭다운 컴포넌트
  *
- * @param defaultValue 초기 선택값 (기본값: 지역 전체)
+ * @param defaultLabel 초기 선택 label (EX: 지역)
  * @param onChange 선택값 변경 시 호출되는 콜백 함수
  * @param className 추가 스타일링을 위한 클래스명
  */
 
 export default function Dropdown({
   options,
-  defaultValue,
+  defaultLabel,
   onChange,
   className,
+  icon = '/icons/chevron-down.svg',
+  selectedValue,
 }: DropdownProps) {
-  const [selectedOption, setSelectedOption] = useState<DropdownOption>(
-    defaultValue || options[0]
-  );
+  const initialOption: DropdownOption = { value: '', label: defaultLabel };
+
+  const [selectedOption, setSelectedOption] =
+    useState<DropdownOption>(initialOption);
   const [isOpen, setIsOpen] = useState<boolean>(false);
+  const dropdownRef = useRef<HTMLDivElement>(null);
+
+  // selectedValue가 변경될 때 selectedOption 업데이트
+  useEffect(() => {
+    if (selectedValue === '') {
+      setSelectedOption(initialOption);
+    } else {
+      const selected =
+        options.find((option) => option.value === selectedValue) ||
+        initialOption;
+      setSelectedOption(selected);
+    }
+  }, [selectedValue, options]);
 
   const handleSelect = (option: DropdownOption) => {
     setSelectedOption(option);
     setIsOpen(false);
     onChange?.(option);
+
+    if (option.value === '') {
+      setSelectedOption(initialOption);
+    }
   };
 
+  const handleClickOutside = (event: MouseEvent) => {
+    if (
+      dropdownRef.current &&
+      !dropdownRef.current.contains(event.target as Node)
+    ) {
+      setIsOpen(false);
+    }
+  };
+
+  useEffect(() => {
+    const listener = (event: MouseEvent) => handleClickOutside(event);
+
+    document.addEventListener('mousedown', listener);
+
+    return () => {
+      document.removeEventListener('mousedown', listener);
+    };
+  }, []);
+
   return (
-    <div className={clsx('relative min-w-[120px]', className)}>
+    <div
+      ref={dropdownRef}
+      className={clsx(
+        'relative',
+        { 'min-w-0 md:max-w-[120px]': icon !== '/icons/chevron-down.svg' },
+        { 'max-w-[120px]': icon === '/icons/chevron-down.svg' }
+      )}
+    >
       <button
         type="button"
         onClick={() => setIsOpen(!isOpen)}
         className={clsx(
-          'h-[42px] w-[120px]',
+          'h-7 md:h-[42px]',
           'relative',
-          'border-2 border-[#808080]',
+          'border-2 border-default-tertiary',
           'rounded-xl',
-          'px-4 py-2',
-          'bg-[#2C2C2C]',
+          'px-3 py-2 md:px-4',
+          'bg-default-tertiary',
           'flex items-center',
-          'gap-[2px]'
+          'gap-2',
+          { 'w-full flex-row-reverse': icon !== '/icons/chevron-down.svg' },
+          className
         )}
       >
-        <span className="flex-1 text-sm font-medium text-white">
+        <span
+          className={`${icon !== '/icons/chevron-down.svg' ? 'hidden md:inline-block' : ''} flex-1 text-xs font-medium text-secondary-40 md:text-sm`}
+        >
           {selectedOption.label}
         </span>
         <Image
-          src="/chevron-down.svg"
+          src={icon}
           width={16}
           height={8}
           alt="dropdown arrow"
@@ -79,21 +131,21 @@ export default function Dropdown({
       {isOpen && (
         <ul
           className={clsx(
-            'absolute left-0 z-10 w-[90px]',
+            'absolute right-0 z-10 w-full max-w-[120px]',
             'mt-2',
             'rounded-xl',
-            'bg-[#404048]',
+            'bg-secondary-80',
             'overflow-hidden shadow-lg'
           )}
         >
           {options.map((option) => (
-            <li key={option.value} className="p-0">
+            <li key={option.value} className="relative p-0">
               <button
                 type="button"
                 onClick={() => handleSelect(option)}
                 className={clsx(
-                  'w-full cursor-pointer px-4 py-2 text-left text-sm',
-                  'text-white hover:bg-[#4A4A52]',
+                  'w-full cursor-pointer px-4 py-2 text-left text-xs md:text-sm',
+                  'text-white-lg rounded-[9px] hover:rounded-[9px] hover:bg-default-primary',
                   'transition-colors duration-150'
                 )}
               >
